add r_dplyr_executor
